src/pyllm/requirements.txt
# LangChain 固定版本
langchain==1.2.0
langchain-community
langchain-openai

# Milvus 依赖
pymilvus==2.3.4

# gRPC
grpcio==1.58.0 # 对应pymilvus==2.3.4
grpcio-tools==1.58.0

# 加载环境变量
python-dotenv

# DL框架--本地运行
# sentence-transformers
# transformers
# torch==2.5.1
